版本升级判断

+++ app/src/main/java/com/community/yuequ/modle/UpgradeInfo.java
@@ -4,7 +4,7 @@ package com.community.yuequ.modle;
  * Created by Administrator on 2016/5/31.
  */
 public class UpgradeInfo {
-    public double version;
+    public float version;
     public String app_path;
     public String is_force_up;//是否强制升级,0：否，1：是
     public String content;//内容条目以#号分割
